Update Dictionary.swift
Using PIPE (Union operator) on dictionaries of the type [String:AnyObject] fails because AnyObject doesn't conform to Equatable.
Change the type constraint of the generic declaration to match Swift's Dictionary generic type.
Now it'll work on any dictionary :]

Didn't test other operators but this one. Chances are they can also be modified in the same way.