return empty string when inverted set range is nil
Strangely, trimming a string that contains only whitespace, results in that same whitespace string, it does *not* result in an empty string. Looking at ltrimmed and rtrimmed, it returns the full string when rangeOfCharacterFromSet returns nil - but when this occurs, it indicates that the string *only* contains characters of the non-inverted set.