Enhance Carthage Support
In a previous change, basic Carthage support was added to the project.
This made the project accessible for Carthage users but only if they
were building for OSX; there were no targets, and no schemes, for iOS.
This change renames the existing OSX targets to ExSwift-OSX and adds
ExSwift-iOS targets to the project.  These targets are for building a
CocoaTouch Framework and the associated tests.  A second shared
scheme, called ExSwift-iOS (the original was renamed ExSwift-OSX) was
created as well.  This will allow Carthage to build for either
platform.  Finally both a CocoaPods and Carthage badge were added to
the README file.